Update Jenkinsfile to remove redundant TAG def
Previously, TAG was defined in the Jenkinsfile and had to be passed to each
script. Local images were built with the tag TAG and the value thus had to be
passed around.

Now only the `push` scripts use the value of TAG, and so it is defined in each
of those scripts. It no longer has to be passed to every script in the Jenkins-
file, and everything running locally relies on the local "latest" tag instead.