Add development channels centralize versions
This adds in edge, latest and branch release channel publishing that
allows downstream projects to set versions to edge, latest, branches or
specific version numbers based on their needs.  Versions used as a base
are lifted to a common bash.env file, and the build process is
consolidated rather than duplicated across multiple build files.

To facilitate this, edge and latest channels are always published, both
internally as well as to dockerhub, while branch channels are only
pushed to the internal registry.