Proposal: Bump ruby to 3.3.1
Hello, I came across your wonderful project and I am curious on trying it out to see if we can use the same
to ubuntu based ruby image compiled with FIPS for a Rails app. I'd like to propose an upgrade to ruby 3.3.1
if that sounds like a good idea, as it would allow me to test it with an existing app :).

Thank you for work